+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 * Rename `Container::modify()` to `Container::set()`
 * Rename `Container::register()` to `Container::add()`
 * Rename `Container::registerAs()` to `Container::copy()`
-* 
+* Rename `Container::unregister()` to `Container::remove()`
 
 ## 0.1.4 (2014.07.07)